requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
pillow==11.2.1
pydantic==2.10.2
PyYAML==6.0.2
requests==2.32.3
tqdm==4.67.1
urllib3==2.4.0
